pyproject.toml
[project]
name = "langgraph_project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.0.20",
    "docx2txt>=0.9",
    "ipykernel>=6.30.1",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.29",
    "langchain-groq>=0.3.7",
    "langchain-openai>=0.3.32",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.6",
    "langsmith>=0.4.19",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "unstructured>=0.18.14",
]
